+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `documentSplitter` field to `Workspace`
+
+### Removed
+- Removed `splitDocuments` boolean field in favour of `documentSplitter` enum field on `Workspace`
 
 ## [6.19.0] - 2023-08-14
 ### Added